Added queue
I made an actual queue for the messages to be written to and read from as well as updating "msg" and "msg_size" to work with it. Also made it able to keep track of how many bytes there are and making sure that the new addition won't make the total number of bytes greater than the BUF_LEN. It checks the size and puts it into bytes_left but tbh idk what to do with that info after it, unless it was only used for the if statement. I'm pushing this separate so you can take a look at it first and see for yourself if it is correct or not/what we are supposed to do, this was all done just based off of my understanding of it so I would be completely understanding if its wrong and will fix it as needed. Lmk what ya think on snap/if anything should be adjusted.